chore(api): split up payloads
* split up payloads into separate files to improve readability
