+++ api/payloads/create_puzzle.go
@@ -1,7 +1,6 @@
 package payloads
 
 import (
-	"errors"
 	"net/http"
 	"time"
 
@@ -13,13 +12,7 @@ import (
 	"github.com/google/uuid"
 )
 
-// Errors
-var (
-	ErrPuzzleInvalidName = errors.New("Invalid puzzle name provided.")
-)
-
 var _ render.Binder = (*CreatePuzzle)(nil)
-var _ render.Binder = (*UpdatePuzzle)(nil)
 
 // CreatePuzzleBlock defines the payload required for creating a user
 type CreatePuzzleBlock struct {
@@ -80,9 +73,10 @@ func (c *CreatePuzzle) ToEntity() entities.Puzzle {
 		Name:        c.Name,
 		Description: sanitize.HTML(c.Description),
 		Difficulty:  c.Difficulty,
-		TimeAllowed: c.TimeAllowed,
 		MaxAttempts: c.MaxAttempts,
-		Groups:      groups,
+		TimeAllowed: c.TimeAllowed,
+
+		Groups: groups,
 	}
 }
 
@@ -90,77 +84,9 @@ func (c *CreatePuzzle) Validate() error {
 	if err := validate.Check(c); err != nil {
 		return internal.NewErrorf(internal.ErrorCodeBadRequest, "%v", err)
 	}
-
 	if err := sanitize.IsClean(c.Name, false); err != nil {
-		return internal.WrapErrorf(err, internal.ErrorCodeBadRequest, "%v", ErrPuzzleInvalidName)
-	}
-
-	return nil
-}
-
-// UpdatePuzzleGroup defines the payload required for updating a puzzle
-type UpdatePuzzleGroup struct {
-	ID          uuid.UUID `json:"id" validate:"required"`
-	Description string    `json:"description" validate:"required,notblank,printasciiextra,max=512"`
-}
-
-// UpdatePuzzle defines the payload required for updating a puzzle
-type UpdatePuzzle struct {
-	Name        string                    `json:"name" validate:"omitempty,notblank,alphanumspace,min=1,max=64"`
-	Description string                    `json:"description" validate:"omitempty,printasciiextra,max=512"`
-	Difficulty  entities.PuzzleDifficulty `json:"difficulty" validate:"omitempty,required,oneof='Easy' 'Medium' 'Hard'"`
-
-	Groups []UpdatePuzzleGroup `json:"groups" validate:"omitempty,dive"`
-}
-
-func (u *UpdatePuzzle) Bind(r *http.Request) error {
-	return nil
-}
-
-func (u *UpdatePuzzle) ToEntity(from entities.Puzzle) entities.Puzzle {
-	update := from
-
-	if len(u.Name) > 0 {
-		update.Name = u.Name
-	}
-	if len(u.Description) > 0 {
-		update.Description = sanitize.HTML(u.Description)
-	}
-	if len(u.Difficulty) > 0 && u.Difficulty != from.Difficulty {
-		update.Difficulty = u.Difficulty
-	}
-	if len(u.Groups) > 0 {
-		groupMap := map[uuid.UUID]UpdatePuzzleGroup{}
-		for _, group := range u.Groups {
-			groupMap[group.ID] = group
-		}
-
-		var groups []entities.PuzzleGroup
-		for _, group := range update.Groups {
-			value, ok := groupMap[group.ID]
-			if !ok {
-				groups = append(groups, group)
-				continue
-			}
-
-			update := group
-			update.Description = sanitize.HTML(value.Description)
-
-			groups = append(groups, update)
-		}
-
-		update.Groups = groups
-	}
-
-	return update
-}
-
-func (u *UpdatePuzzle) Validate() error {
-	if err := validate.Check(u); err != nil {
 		return internal.NewErrorf(internal.ErrorCodeBadRequest, "%v", err)
 	}
-	if err := sanitize.IsClean(u.Name, false); err != nil {
-		return internal.WrapErrorf(err, internal.ErrorCodeBadRequest, "%v", ErrPuzzleInvalidName)
-	}
+
 	return nil
 }
